Remove ScratchViewsNGP.
UnitTestScratchViews.C now uses ScratchViews, which has absorbed
the necessary changes from ScratchViewsNGP in order to operate on
the GPU.

Calls to master-element-related functions are still ifdef'd out for the
GPU.

fill_pre_req_data is overloaded, such that it gathers field data from
stk-fields if ElemDataRequestsNGP is passed in, but gathers from
ngp::Fields if ElemDataRequestsGPU is passed in. I think that ultimately
we will get rid of ElemDataRequestsNGP and then quite a bit of the
near-duplicate code can be removed.

This commit also GPU-izes SharedMemData (but not
SharedMemData_faceElem). UnitTestScratchViews.C now contains a test
of SharedMemData, and it runs on GPU.
