+++ include/KokkosInterface.h
@@ -101,6 +101,7 @@ SharedMemView<T***,TeamShmemType> get_shmem_view_3D(const TEAMHANDLETYPE& team, 
   return Kokkos::subview(SharedMemView<T****,TeamShmemType>(team.team_scratch(1), team.team_size(), len1, len2, len3), team.team_rank(), Kokkos::ALL(), Kokkos::ALL(), Kokkos::ALL());
 }
 
+#ifndef KOKKOS_ENABLE_CUDA
 template<typename T, typename TEAMHANDLETYPE, typename TeamShmemType=HostShmem>
 KOKKOS_FUNCTION
 SharedMemView<T****,TeamShmemType> get_shmem_view_4D(const TEAMHANDLETYPE& team, size_t len1, size_t len2, size_t len3, size_t len4)
@@ -114,6 +115,7 @@ SharedMemView<T*****,TeamShmemType> get_shmem_view_5D(const TEAMHANDLETYPE& team
 {
   return Kokkos::subview(SharedMemView<T******,TeamShmemType>(team.team_scratch(1), team.team_size(), len1, len2, len3, len4, len5), team.team_rank(), Kokkos::ALL(), Kokkos::ALL(), Kokkos::ALL());
 }
+#endif
 
 template<typename SizeType, class Function>
 void kokkos_parallel_for(const std::string& debuggingName, SizeType n, Function loop_body)
