+++ include/SharedMemData.h
@@ -20,41 +20,104 @@
 namespace sierra{
 namespace nalu{
 
+template<typename TEAMHANDLETYPE, typename SHMEM>
 struct SharedMemData {
-    SharedMemData(const sierra::nalu::TeamHandleType& team,
+    KOKKOS_FUNCTION
+    SharedMemData(const TEAMHANDLETYPE& team,
+         unsigned nDim,
+         const ElemDataRequestsGPU& dataNeededByKernels,
+         unsigned nodesPerEntity,
+         unsigned rhsSize)
+     : simdPrereqData(team, nDim, nodesPerEntity, dataNeededByKernels)
+#ifdef KOKKOS_ENABLE_CUDA
+      ,prereqData_byValue(team, nDim, nodesPerEntity, dataNeededByKernels)
+#endif
+    {
+#ifndef KOKKOS_ENABLE_CUDA
+        for(int simdIndex=0; simdIndex<simdLen; ++simdIndex) {
+          prereqData[simdIndex] = std::unique_ptr<ScratchViews<double,TEAMHANDLETYPE,SHMEM> >(new ScratchViews<double,TEAMHANDLETYPE,SHMEM>(team, nDim, nodesPerEntity, dataNeededByKernels));
+        }
+#else
+        prereqData[0] = &prereqData_byValue;
+#endif
+        simdrhs = get_shmem_view_1D<DoubleType,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        simdlhs = get_shmem_view_2D<DoubleType,TEAMHANDLETYPE,SHMEM>(team, rhsSize, rhsSize);
+        rhs = get_shmem_view_1D<double,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        lhs = get_shmem_view_2D<double,TEAMHANDLETYPE,SHMEM>(team, rhsSize, rhsSize);
+
+        scratchIds = get_shmem_view_1D<int,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        sortPermutation = get_shmem_view_1D<int,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+    }
+
+    SharedMemData(const TEAMHANDLETYPE& team,
          unsigned nDim,
          const ElemDataRequestsNGP& dataNeededByKernels,
          unsigned nodesPerEntity,
          unsigned rhsSize)
      : simdPrereqData(team, nDim, nodesPerEntity, dataNeededByKernels)
+#ifdef KOKKOS_ENABLE_CUDA
+      ,prereqData_byValue(team, nDim, nodesPerEntity, dataNeededByKernels)
+#endif
     {
+#ifndef KOKKOS_ENABLE_CUDA
         for(int simdIndex=0; simdIndex<simdLen; ++simdIndex) {
-          prereqData[simdIndex] = std::unique_ptr<ScratchViews<double> >(new ScratchViews<double>(team, nDim, nodesPerEntity, dataNeededByKernels));
+          prereqData[simdIndex] = std::unique_ptr<ScratchViews<double,TEAMHANDLETYPE,SHMEM> >(new ScratchViews<double,TEAMHANDLETYPE,SHMEM>(team, nDim, nodesPerEntity, dataNeededByKernels));
         }
-        simdrhs = get_shmem_view_1D<DoubleType>(team, rhsSize);
-        simdlhs = get_shmem_view_2D<DoubleType>(team, rhsSize, rhsSize);
-        rhs = get_shmem_view_1D<double>(team, rhsSize);
-        lhs = get_shmem_view_2D<double>(team, rhsSize, rhsSize);
+#else
+        prereqData[0] = &prereqData_byValue;
+#endif
+        simdrhs = get_shmem_view_1D<DoubleType,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        simdlhs = get_shmem_view_2D<DoubleType,TEAMHANDLETYPE,SHMEM>(team, rhsSize, rhsSize);
+        rhs = get_shmem_view_1D<double,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        lhs = get_shmem_view_2D<double,TEAMHANDLETYPE,SHMEM>(team, rhsSize, rhsSize);
 
-        scratchIds = get_shmem_view_1D<int>(team, rhsSize);
-        sortPermutation = get_shmem_view_1D<int>(team, rhsSize);
+        scratchIds = get_shmem_view_1D<int,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        sortPermutation = get_shmem_view_1D<int,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
     }
 
     const stk::mesh::Entity* elemNodes[simdLen];
     int numSimdElems;
-    std::unique_ptr<ScratchViews<double>> prereqData[simdLen];
-    ScratchViews<DoubleType> simdPrereqData;
-    SharedMemView<DoubleType*> simdrhs;
-    SharedMemView<DoubleType**> simdlhs;
-    SharedMemView<double*> rhs;
-    SharedMemView<double**> lhs;
+#ifdef KOKKOS_ENABLE_CUDA
+    ScratchViews<double,TEAMHANDLETYPE,SHMEM> prereqData_byValue;
+    ScratchViews<double,TEAMHANDLETYPE,SHMEM>* prereqData[1];
+#else
+    std::unique_ptr<ScratchViews<double,TEAMHANDLETYPE,SHMEM>> prereqData[simdLen];
+#endif
+    ScratchViews<DoubleType,TEAMHANDLETYPE,SHMEM> simdPrereqData;
+    SharedMemView<DoubleType*,SHMEM> simdrhs;
+    SharedMemView<DoubleType**,SHMEM> simdlhs;
+    SharedMemView<double*,SHMEM> rhs;
+    SharedMemView<double**,SHMEM> lhs;
 
-    SharedMemView<int*> scratchIds;
-    SharedMemView<int*> sortPermutation;
+    SharedMemView<int*,SHMEM> scratchIds;
+    SharedMemView<int*,SHMEM> sortPermutation;
 };
 
+template<typename TEAMHANDLETYPE, typename SHMEM>
 struct SharedMemData_FaceElem {
-    SharedMemData_FaceElem(const sierra::nalu::TeamHandleType& team,
+    SharedMemData_FaceElem(const TEAMHANDLETYPE& team,
+         unsigned nDim,
+         const ElemDataRequestsGPU& faceDataNeeded,
+         const ElemDataRequestsGPU& elemDataNeeded,
+         const ScratchMeInfo& meElemInfo,
+         unsigned rhsSize)
+     : simdFaceViews(team, nDim, meElemInfo.nodesPerFace_, faceDataNeeded),
+       simdElemViews(team, nDim, meElemInfo, elemDataNeeded)
+    {
+        for(int simdIndex=0; simdIndex<simdLen; ++simdIndex) {
+          faceViews[simdIndex] = std::unique_ptr<ScratchViews<double,TEAMHANDLETYPE,SHMEM> >(new ScratchViews<double,TEAMHANDLETYPE,SHMEM>(team, nDim, meElemInfo.nodesPerFace_, faceDataNeeded));
+          elemViews[simdIndex] = std::unique_ptr<ScratchViews<double,TEAMHANDLETYPE,SHMEM> >(new ScratchViews<double,TEAMHANDLETYPE,SHMEM>(team, nDim, meElemInfo, elemDataNeeded));
+        }
+        simdrhs = get_shmem_view_1D<DoubleType,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        simdlhs = get_shmem_view_2D<DoubleType,TEAMHANDLETYPE,SHMEM>(team, rhsSize, rhsSize);
+        rhs = get_shmem_view_1D<double,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        lhs = get_shmem_view_2D<double,TEAMHANDLETYPE,SHMEM>(team, rhsSize, rhsSize);
+
+        scratchIds = get_shmem_view_1D<int,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        sortPermutation = get_shmem_view_1D<int,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+    }
+
+    SharedMemData_FaceElem(const TEAMHANDLETYPE& team,
          unsigned nDim,
          const ElemDataRequestsNGP& faceDataNeeded,
          const ElemDataRequestsNGP& elemDataNeeded,
@@ -64,32 +127,32 @@ struct SharedMemData_FaceElem {
        simdElemViews(team, nDim, meElemInfo, elemDataNeeded)
     {
         for(int simdIndex=0; simdIndex<simdLen; ++simdIndex) {
-          faceViews[simdIndex] = std::unique_ptr<ScratchViews<double> >(new ScratchViews<double>(team, nDim, meElemInfo.nodesPerFace_, faceDataNeeded));
-          elemViews[simdIndex] = std::unique_ptr<ScratchViews<double> >(new ScratchViews<double>(team, nDim, meElemInfo, elemDataNeeded));
+          faceViews[simdIndex] = std::unique_ptr<ScratchViews<double,TEAMHANDLETYPE,SHMEM> >(new ScratchViews<double,TEAMHANDLETYPE,SHMEM>(team, nDim, meElemInfo.nodesPerFace_, faceDataNeeded));
+          elemViews[simdIndex] = std::unique_ptr<ScratchViews<double,TEAMHANDLETYPE,SHMEM> >(new ScratchViews<double,TEAMHANDLETYPE,SHMEM>(team, nDim, meElemInfo, elemDataNeeded));
         }
-        simdrhs = get_shmem_view_1D<DoubleType>(team, rhsSize);
-        simdlhs = get_shmem_view_2D<DoubleType>(team, rhsSize, rhsSize);
-        rhs = get_shmem_view_1D<double>(team, rhsSize);
-        lhs = get_shmem_view_2D<double>(team, rhsSize, rhsSize);
+        simdrhs = get_shmem_view_1D<DoubleType,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        simdlhs = get_shmem_view_2D<DoubleType,TEAMHANDLETYPE,SHMEM>(team, rhsSize, rhsSize);
+        rhs = get_shmem_view_1D<double,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        lhs = get_shmem_view_2D<double,TEAMHANDLETYPE,SHMEM>(team, rhsSize, rhsSize);
 
-        scratchIds = get_shmem_view_1D<int>(team, rhsSize);
-        sortPermutation = get_shmem_view_1D<int>(team, rhsSize);
+        scratchIds = get_shmem_view_1D<int,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
+        sortPermutation = get_shmem_view_1D<int,TEAMHANDLETYPE,SHMEM>(team, rhsSize);
     }
 
     const stk::mesh::Entity* connectedNodes[simdLen];
     int numSimdFaces;
     int elemFaceOrdinal;
-    std::unique_ptr<ScratchViews<double>> faceViews[simdLen];
-    std::unique_ptr<ScratchViews<double>> elemViews[simdLen];
-    ScratchViews<DoubleType> simdFaceViews;
-    ScratchViews<DoubleType> simdElemViews;
-    SharedMemView<DoubleType*> simdrhs;
-    SharedMemView<DoubleType**> simdlhs;
-    SharedMemView<double*> rhs;
-    SharedMemView<double**> lhs;
+    std::unique_ptr<ScratchViews<double,TEAMHANDLETYPE,SHMEM>> faceViews[simdLen];
+    std::unique_ptr<ScratchViews<double,TEAMHANDLETYPE,SHMEM>> elemViews[simdLen];
+    ScratchViews<DoubleType,TEAMHANDLETYPE,SHMEM> simdFaceViews;
+    ScratchViews<DoubleType,TEAMHANDLETYPE,SHMEM> simdElemViews;
+    SharedMemView<DoubleType*,SHMEM> simdrhs;
+    SharedMemView<DoubleType**,SHMEM> simdlhs;
+    SharedMemView<double*,SHMEM> rhs;
+    SharedMemView<double**,SHMEM> lhs;
 
-    SharedMemView<int*> scratchIds;
-    SharedMemView<int*> sortPermutation;
+    SharedMemView<int*,SHMEM> scratchIds;
+    SharedMemView<int*,SHMEM> sortPermutation;
 };
 
 } // namespace nalu
