+++ unit_tests/UnitTestScratchViews.C
@@ -9,7 +9,7 @@
 
 #include <SimdInterface.h>
 #include <ElemDataRequestsGPU.h>
-#include <ScratchViewsNGP.h>
+#include <ScratchViews.h>
 #include <kernel/Kernel.h>
 
 #include <stk_ngp/Ngp.hpp>
@@ -28,7 +28,16 @@ class TestKernel
   void execute(
     sierra::nalu::SharedMemView<double**,ShmemType> &lhs,
     sierra::nalu::SharedMemView<double*,ShmemType> &rhs,
-    sierra::nalu::ScratchViewsNGP<double> &scratchViews) const
+    sierra::nalu::ScratchViews<sierra::nalu::DoubleType,TeamType,ShmemType> &scratchViews) const
+  {
+    printf("TestKernel::execute!!\n");
+  }
+
+  KOKKOS_FUNCTION
+  void execute(
+    sierra::nalu::SharedMemView<double**,ShmemType> &lhs,
+    sierra::nalu::SharedMemView<double*,ShmemType> &rhs,
+    sierra::nalu::ScratchViews<double,TeamType,ShmemType> &scratchViews) const
   {
     printf("TestKernel::execute!!\n");
   }
@@ -89,7 +98,7 @@ void do_the_test(stk::mesh::BulkData& bulk, sierra::nalu::ScalarFieldType* press
   {
     const ngp::Mesh::BucketType& b = ngpMesh.get_bucket(stk::topology::ELEM_RANK, team.league_rank());
 
-    sierra::nalu::ScratchViewsNGP<double> scrviews(team, ngpMesh.get_spatial_dimension(), numNodes, dataNGP);
+    sierra::nalu::ScratchViews<double,TeamType,ShmemType> scrviews(team, ngpMesh.get_spatial_dimension(), numNodes, dataNGP);
 
     sierra::nalu::SharedMemView<double**,ShmemType> simdlhs =
         sierra::nalu::get_shmem_view_2D<double,TeamType,ShmemType>(team, rhsSize, rhsSize);
@@ -130,6 +139,99 @@ TEST_F(Hex8MeshWithNSOFields, ScratchViews)
   fill_mesh_and_initialize_test_fields("generated:2x2x2");
 
   do_the_test(bulk, pressure, velocity);
+}
+
+void do_the_smdata_test(stk::mesh::BulkData& bulk, sierra::nalu::ScalarFieldType* pressure, sierra::nalu::VectorFieldType* velocity)
+{
+  stk::topology elemTopo = stk::topology::HEX_8;
+  sierra::nalu::ElemDataRequests dataReq;
+  auto meSCV = sierra::nalu::MasterElementRepo::get_volume_master_element(elemTopo);
+  dataReq.add_cvfem_volume_me(meSCV);
+
+  dataReq.add_gathered_nodal_field(*velocity, 3);
+  dataReq.add_gathered_nodal_field(*pressure, 1);
+
+  EXPECT_EQ(2u, dataReq.get_fields().size());
+
+  const stk::mesh::MetaData& meta = bulk.mesh_meta_data();
+
+  sierra::nalu::ElemDataRequestsGPU dataNGP(dataReq, meta.get_fields().size());
+
+  const int numNodes = elemTopo.num_nodes();
+  const int rhsSize = numNodes;
+  const int lhsSize = rhsSize * rhsSize;
+
+  const unsigned velocityOrdinal = velocity->mesh_meta_data_ordinal();
+  const unsigned pressureOrdinal = pressure->mesh_meta_data_ordinal();
+
+  const int bytes_per_team = 0;
+  const int bytes_per_thread =
+    (sierra::nalu::calculate_shared_mem_bytes_per_thread(
+       lhsSize, rhsSize, rhsSize, meta.spatial_dimension(), dataNGP) +
+     (rhsSize + lhsSize) * sizeof(double) * sierra::nalu::simdLen);
+
+  const bool interleaveMEViews = false;
+
+  int numResults = 5;
+  IntViewType result("result", numResults);
+
+  Kokkos::deep_copy(result.h_view, 0);
+  result.template modify<typename IntViewType::host_mirror_space>();
+  result.template sync<typename IntViewType::execution_space>();
+
+  ngp::Mesh ngpMesh(bulk);
+
+  TestKernel testKernel(velocityOrdinal, pressureOrdinal);
+
+  int threads_per_team = 1;
+  auto team_exec = sierra::nalu::get_device_team_policy(ngpMesh.num_buckets(stk::topology::ELEM_RANK), bytes_per_team, bytes_per_thread, threads_per_team);
+
+  Kokkos::parallel_for(team_exec, KOKKOS_LAMBDA(const sierra::nalu::DeviceTeamHandleType& team)
+  {
+    const ngp::Mesh::BucketType& b = ngpMesh.get_bucket(stk::topology::ELEM_RANK, team.league_rank());
+
+    sierra::nalu::SharedMemData<TeamType,ShmemType> smdata(team, ngpMesh.get_spatial_dimension(), dataNGP, numNodes, rhsSize);
+
+    NGP_ThrowAssert(smdata.simdPrereqData.total_bytes() != 0);
+    const size_t bucketLen   = b.size();
+ 
+    Kokkos::parallel_for(Kokkos::TeamThreadRange(team, bucketLen), [&](const size_t& bktIndex)
+    {
+        stk::mesh::Entity element = b[bktIndex];
+        sierra::nalu::fill_pre_req_data(dataNGP, ngpMesh, stk::topology::ELEM_RANK, element,
+                          *smdata.prereqData[0], interleaveMEViews);
+
+#ifndef KOKKOS_ENABLE_CUDA
+//no copy-interleave needed on GPU since no simd.
+        sierra::nalu::copy_and_interleave(smdata.prereqData, 1, smdata.simdPrereqData, interleaveMEViews);
+#endif
+
+        auto& velocityView = smdata.prereqData[0]->get_scratch_view_2D(velocityOrdinal);
+        auto& pressureView = smdata.prereqData[0]->get_scratch_view_1D(pressureOrdinal);
+
+        result.d_view(0) = (pressureView(0) - 1.0) < 1.e-9 ? 1 : 0;
+        result.d_view(1) = (pressureView(7) - 1.0) < 1.e-9 ? 1 : 0;
+        result.d_view(2) = (velocityView(6,0) - 1.0) < 1.e-9 ? 1 : 0;
+        result.d_view(3) = (velocityView(6,1) - 0.0) < 1.e-9 ? 1 : 0;
+        result.d_view(4) = (velocityView(6,2) - 0.0) < 1.e-9 ? 1 : 0;
+
+        testKernel.execute(smdata.lhs, smdata.rhs, smdata.simdPrereqData);
+    });
+  });
+
+  result.modify<IntViewType::execution_space>();
+  result.sync<IntViewType::host_mirror_space>();
+
+  for(int i=0; i<numResults; ++i) {
+    EXPECT_EQ(1, result.h_view(i));
+  }
+}
+
+TEST_F(Hex8MeshWithNSOFields, SharedMemData)
+{
+  fill_mesh_and_initialize_test_fields("generated:2x2x2");
+
+  do_the_smdata_test(bulk, pressure, velocity);
 }
 
 #ifdef KOKKOS_HAVE_CUDA
