+++ unit_tests/kernels/UnitTestKernelUtils.C
@@ -585,7 +585,7 @@ void calc_mass_flow_rate_scs(
   sierra::nalu::ElemDataRequestsNGP dataNeededNGP(dataNeeded, meta.get_fields().size());
 
   const int bytes_per_team = 0;
-  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data(
+  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data<double>(
     dataNeededNGP, meta.spatial_dimension()) ;
 
   auto team_exec = sierra::nalu::get_host_team_policy(
@@ -667,7 +667,7 @@ void calc_projected_nodal_gradient_interior(
   sierra::nalu::ElemDataRequestsNGP dataNeededNGP(dataNeeded, meta.get_fields().size());
 
   const int bytes_per_team = 0;
-  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data(dataNeededNGP, meta.spatial_dimension()) ;
+  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data<double>(dataNeededNGP, meta.spatial_dimension()) ;
 
   auto v_shape_function = Kokkos::View<double**>("shape_function", meSCS->numIntPoints_, meSCS->nodesPerElement_);
 
@@ -691,7 +691,7 @@ void calc_projected_nodal_gradient_interior(
       auto v_dnv = preReqData.get_scratch_view_1D(dnv);
       auto v_scalar = preReqData.get_scratch_view_1D(scalarField);
       auto v_scs_areav = preReqData.get_me_views(sierra::nalu::CURRENT_COORDINATES).scs_areav;
-      const stk::mesh::Entity* node_rels = preReqData.elemNodes;
+      const ngp::Mesh::ConnectedNodes node_rels = preReqData.elemNodes;
       const int* lrscv = meSCS->adjacentNodes();
 
       for (int ip = 0; ip < meSCS->numIntPoints_; ++ip) {
@@ -746,7 +746,7 @@ void calc_projected_nodal_gradient_interior(
   sierra::nalu::ElemDataRequestsNGP dataNeededNGP(dataNeeded, meta.get_fields().size());
 
   const int bytes_per_team = 0;
-  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data(dataNeededNGP, meta.spatial_dimension()) ;
+  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data<double>(dataNeededNGP, meta.spatial_dimension()) ;
 
   auto v_shape_function = Kokkos::View<double**>("shape_function", meSCS->numIntPoints_, meSCS->nodesPerElement_);
 
@@ -770,7 +770,7 @@ void calc_projected_nodal_gradient_interior(
       auto v_dnv = preReqData.get_scratch_view_1D(dnv);
       auto v_vector = preReqData.get_scratch_view_2D(vectorField);
       auto v_scs_areav = preReqData.get_me_views(sierra::nalu::CURRENT_COORDINATES).scs_areav;
-      const stk::mesh::Entity* node_rels = preReqData.elemNodes;
+      const ngp::Mesh::ConnectedNodes node_rels = preReqData.elemNodes;
       const int* lrscv = meSCS->adjacentNodes();
 
       for (int di = 0; di < ndim; ++di) {
@@ -828,7 +828,7 @@ void calc_projected_nodal_gradient_boundary(
   sierra::nalu::ElemDataRequestsNGP dataNeededNGP(dataNeeded, meta.get_fields().size());
 
   const int bytes_per_team = 0;
-  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data(dataNeededNGP, meta.spatial_dimension()) ;
+  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data<double>(dataNeededNGP, meta.spatial_dimension()) ;
 
   auto v_shape_function = Kokkos::View<double**>("shape_function", meBC->numIntPoints_, meBC->nodesPerElement_);
 
@@ -851,7 +851,7 @@ void calc_projected_nodal_gradient_boundary(
       auto v_dnv = preReqData.get_scratch_view_1D(dnv);
       auto v_scalar = preReqData.get_scratch_view_1D(scalarField);
       auto v_scs_areav = preReqData.get_me_views(sierra::nalu::CURRENT_COORDINATES).scs_areav;
-      const stk::mesh::Entity* node_rels = preReqData.elemNodes;
+      const ngp::Mesh::ConnectedNodes node_rels = preReqData.elemNodes;
       const int* ipNodeMap = meBC->ipNodeMap();
 
       for (int ip = 0; ip < meBC->numIntPoints_; ++ip) {
@@ -902,7 +902,7 @@ void calc_projected_nodal_gradient_boundary(
   sierra::nalu::ElemDataRequestsNGP dataNeededNGP(dataNeeded, meta.get_fields().size());
 
   const int bytes_per_team = 0;
-  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data(dataNeededNGP, meta.spatial_dimension()) ;
+  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data<double>(dataNeededNGP, meta.spatial_dimension()) ;
 
   auto v_shape_function = Kokkos::View<double**>("shape_function", meBC->numIntPoints_, meBC->nodesPerElement_);
 
@@ -926,7 +926,7 @@ void calc_projected_nodal_gradient_boundary(
       auto v_dnv = preReqData.get_scratch_view_1D(dnv);
       auto v_vector = preReqData.get_scratch_view_2D(vectorField);
       auto v_scs_areav = preReqData.get_me_views(sierra::nalu::CURRENT_COORDINATES).scs_areav;
-      const stk::mesh::Entity* node_rels = preReqData.elemNodes;
+      const ngp::Mesh::ConnectedNodes node_rels = preReqData.elemNodes;
       const int* ipNodeMap = meBC->ipNodeMap();
 
       for (int di = 0; di < ndim; ++di) {
@@ -973,7 +973,7 @@ void calc_dual_nodal_volume(
   sierra::nalu::ElemDataRequestsNGP dataNeededNGP(dataNeeded, meta.get_fields().size());
 
   const int bytes_per_team = 0;
-  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data(dataNeededNGP, meta.spatial_dimension()) ;
+  const int bytes_per_thread = sierra::nalu::get_num_bytes_pre_req_data<double>(dataNeededNGP, meta.spatial_dimension()) ;
 
   auto v_shape_function = Kokkos::View<double**>("shape_function", meSCV->numIntPoints_, meSCV->nodesPerElement_);
 
@@ -994,7 +994,7 @@ void calc_dual_nodal_volume(
       sierra::nalu::fill_pre_req_data(dataNeededNGP, bulk, element, preReqData);
 
       auto v_scv_vol = preReqData.get_me_views(sierra::nalu::CURRENT_COORDINATES).scv_volume;
-      const stk::mesh::Entity* node_rels = preReqData.elemNodes;
+      const ngp::Mesh::ConnectedNodes node_rels = preReqData.elemNodes;
       const int* ipNodeMap = meSCV->ipNodeMap();
 
       for (int ip = 0; ip < meSCV->numIntPoints_; ++ip) {
